+++ lib/src/gatekeeper_client.dart
@@ -158,14 +158,14 @@ class GatekeeperClient {
   ///
   /// - [accessKey]: The access key used to authenticate the login.
   ///
-  /// Returns a [Future] that completes with `true` if login was successful,
-  /// or `false` if it failed.
-  Future<bool> login(String accessKey) async {
+  /// Returns a [Future] that completes with `(ok: true, serverVersion: 'x.x.x'})` if login was successful,
+  /// or `(ok: false, serverVersion: null})` if it failed.
+  Future<({bool ok, String? serverVersion})> login(String accessKey) async {
     _accessKey = accessKey;
 
     if (secure) {
       var ok = await _exchangeSessionKey();
-      if (!ok) return false;
+      if (!ok) return (ok: false, serverVersion: null);
     }
 
     var sessionKey = chainAESEncryptor.sessionKey;
@@ -174,16 +174,25 @@ class GatekeeperClient {
     var accessKeyBase64 = base64.encode(accessKeyHash);
 
     var response = await _sendCommand("login $accessKeyBase64");
-    var logged = response?.contains('true') ?? false;
-    if (logged) {
-      _logged = true;
+
+    var logged = false;
+    String? serverVersion;
+
+    if (response != null && response.isNotEmpty) {
+      logged = response.contains('true');
+      if (logged) {
+        _logged = true;
+      }
+
+      serverVersion = RegExp(r'\[(.*?)]').firstMatch(response)?.group(1);
     }
 
     if (verbose) {
-      print('-- LOGIN: $logged');
+      print(
+          '-- LOGIN: $logged ${serverVersion != null ? '[$serverVersion]' : ''}');
     }
 
-    return logged;
+    return (ok: true, serverVersion: serverVersion);
   }
 
   Future<bool> _exchangeSessionKey() async {